aider: Disabled pinch-to-zoom feature on iPad's Safari browser by adding a JavaScript event listener to prevent the default behavior.
# Aider chat conversation:

USER: I can still pinch to zoom

ASSISTANT: I apologize for the inconvenience. The `maximum-scale=1.0, user-scalable=no` in the viewport meta tag should have disabled the pinch-to-zoom feature. However, it seems that this is not working as expected on your iPad's Safari browser.

Unfortunately, as of iOS 10, Safari on iOS ignores the `maximum-scale=1.0, user-scalable=no` settings in the viewport meta tag due to accessibility reasons. This means that users can still pinch-to-zoom on a webpage even if it's disabled in the viewport meta tag.

A workaround to this issue is to use a JavaScript event listener to prevent the default pinch-to-zoom behavior. Here is the change:

```html
index.html
<<<<<<< HEAD
</body>
</html>
=======
    <script>
        document.addEventListener('touchmove', function (event) {
            if (event.scale !== 1) { event.preventDefault(); }
        }, { passive: false });
    </script>
</body>
</html>
>>>>>>> updated
```

This JavaScript code listens for the `touchmove` event and prevents the default behavior if the scale is not 1, effectively disabling pinch-to-zoom. Please note that this might not work in all cases as some browsers might not support this feature or have it disabled for accessibility reasons.