aider: Set isProcessing to false after hiding matched cards.
# Aider chat conversation:

USER: After the first matched pair, I still can't tap a card to continue the game. After a matched pair disappears, please make sure all of the variables are set such that the game can continue.

ASSISTANT: I see. It seems like the issue might be related to the `isProcessing` variable. This variable is used to prevent the player from flipping more than two cards at a time. However, it seems like it's not being reset to `false` after a pair of cards is matched and hidden, which is preventing the player from flipping any more cards.

To fix this, we need to ensure that `isProcessing` is set to `false` after the matched cards are hidden. This can be done by adding `isProcessing = false;` at the end of the `hideMatchedCards` function.

Here is the change:

```html
index.html
<<<<<<< HEAD
        // Function to hide matched cards
        function hideMatchedCards() {
            const matchedCards = document.querySelectorAll('.card.matched');
            matchedCards.forEach(card => {
                card.style.backgroundColor = 'transparent';
                card.style.border = 'none';
                card.removeEventListener('click', flipCard);
            });
        }
=======
        // Function to hide matched cards
        function hideMatchedCards() {
            const matchedCards = document.querySelectorAll('.card.matched');
            matchedCards.forEach(card => {
                card.style.backgroundColor = 'transparent';
                card.style.border = 'none';
                card.removeEventListener('click', flipCard);
            });
            isProcessing = false;
        }
>>>>>>> updated
```

This change will ensure that the game can continue after a pair of cards is matched and hidden.